Move UMAPINFO functionality to g_level.cpp; remove unnecessary .cpp/.h files
