--- common/umapinfo.h
@@ -1,26 +1,0 @@
-//-----------------------------------------------------------------------------
-//
-// Copyright (C) 2017 Christoph Oelckers
-// Copyright (C) 2006-2020 by The Odamex Team.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 2 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see http://www.gnu.org/licenses/
-//
-//-----------------------------------------------------------------------------
-
-#ifndef __UMAPINFO_H
-#define __UMAPINFO_H
-
-void ParseUMapInfoLump(int lump, const char* lumpname);
-
-#endif
